Fix Db Connection Error

+++ monitoring_server/src/main/resources/db/schema.sql
@@ -14,9 +14,9 @@ CREATE TABLE IF NOT EXISTS system_mon.system_status_daily (
     id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     hostname VARCHAR(50) NOT NULL,
     cpu_avg DOUBLE PRECISION NOT NULL,                 -- double -> DOUBLE PRECISION (정밀한 평균값)
-    cpu_max REAL NOT NULL,                             -- 하루 중 피크 타임 맥스값
+    cpu_max DOUBLE PRECISION NOT NULL,                             -- 하루 중 피크 타임 맥스값
     memory_avg DOUBLE PRECISION NOT NULL,
-    memory_max REAL NOT NULL,
+    memory_max DOUBLE PRECISION NOT NULL,
     network_rx_total BIGINT NOT NULL,                  -- 하루 동안의 총 전송량
     network_tx_total BIGINT NOT NULL,
     log_date DATE NOT NULL,
